Moved crd checker to kubernetes-routes
